+++ app/src/main/java/xyz/thingapps/rssliveslider/api/RssApi.kt
@@ -3,6 +3,7 @@ package xyz.thingapps.rssliveslider.api
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Url
+import xyz.thingapps.rssliveslider.models.Cast
 
 interface RssApi {
     @GET
