+++ app/src/main/java/xyz/thingapps/rssliveslider/models/Media.kt
@@ -1,0 +1,36 @@
+package xyz.thingapps.rssliveslider.models
+
+import android.os.Parcel
+import android.os.Parcelable
+import com.tickaroo.tikxml.annotation.Attribute
+import com.tickaroo.tikxml.annotation.Xml
+
+@Xml(name = "enclosure")
+data class Media(
+    @Attribute val url: String = "",
+    @Attribute val type: String = ""
+) : Parcelable {
+    constructor(parcel: Parcel) : this(
+        parcel.readString() ?: "",
+        parcel.readString() ?: ""
+    )
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeString(url)
+        parcel.writeString(type)
+    }
+
+    override fun describeContents(): Int {
+        return 0
+    }
+
+    companion object CREATOR : Parcelable.Creator<Media> {
+        override fun createFromParcel(parcel: Parcel): Media {
+            return Media(parcel)
+        }
+
+        override fun newArray(size: Int): Array<Media?> {
+            return arrayOfNulls(size)
+        }
+    }
+}
