+++ app/src/main/java/xyz/thingapps/rssliveslider/models/RssUrl.kt
@@ -1,0 +1,3 @@
+package xyz.thingapps.rssliveslider.models
+
+data class RssUrl(val url: String, val createdAt: Long)
